image-scanning-code/requirements.txt
flask
waitress
azure-cli
azure-identity
azure-mgmt-resourcegraph
python-dotenv
pytest
pika